refactor: reorganize contract versioning code into dedicated contracts package
Move all contract-related commands, queries, model classes, and utilities
from scattered locations into a cohesive templates/contracts/ package with
commands/, queries/, and model/ sub-packages. Rename
TemplateContractCompatibilityService.kt to TemplateCompatibilityResult.kt.
