+++ apps/epistola/src/test/kotlin/app/epistola/suite/templates/DocumentTemplateRoutesTest.kt
@@ -6,10 +6,10 @@ import app.epistola.suite.common.ids.CatalogId
 import app.epistola.suite.common.ids.TemplateId
 import app.epistola.suite.common.ids.TenantId
 import app.epistola.suite.common.ids.VariantKey
+import app.epistola.suite.templates.contracts.queries.GetLatestContractVersion
 import app.epistola.suite.templates.model.DataExample
 import app.epistola.suite.templates.queries.ListDocumentTemplates
 import app.epistola.suite.templates.queries.ListDocumentTemplatesHandler
-import app.epistola.suite.templates.queries.contracts.GetLatestContractVersion
 import app.epistola.suite.tenants.Tenant
 import org.assertj.core.api.Assertions.assertThat
 import org.jdbi.v3.core.Jdbi
