+++ modules/epistola-core/src/main/kotlin/app/epistola/suite/templates/analysis/TemplateCompatibilityResult.kt
