+++ modules/epistola-core/src/main/kotlin/app/epistola/suite/templates/contracts/commands/UpdateContractVersion.kt
@@ -1,4 +1,4 @@
-package app.epistola.suite.templates.commands.contracts
+package app.epistola.suite.templates.contracts.commands
 
 import app.epistola.suite.catalog.requireCatalogEditable
 import app.epistola.suite.common.ids.TemplateId
@@ -7,7 +7,7 @@ import app.epistola.suite.mediator.Command
 import app.epistola.suite.mediator.CommandHandler
 import app.epistola.suite.security.Permission
 import app.epistola.suite.security.RequiresPermission
-import app.epistola.suite.templates.model.ContractVersion
+import app.epistola.suite.templates.contracts.model.ContractVersion
 import app.epistola.suite.templates.model.DataExample
 import app.epistola.suite.templates.validation.JsonSchemaValidator
 import org.jdbi.v3.core.Jdbi
