+++ modules/epistola-core/src/main/kotlin/app/epistola/suite/templates/contracts/queries/GetLatestContractVersion.kt
@@ -1,4 +1,4 @@
-package app.epistola.suite.templates.queries.contracts
+package app.epistola.suite.templates.contracts.queries
 
 import app.epistola.suite.common.ids.TemplateId
 import app.epistola.suite.common.ids.TenantKey
@@ -6,17 +6,15 @@ import app.epistola.suite.mediator.Query
 import app.epistola.suite.mediator.QueryHandler
 import app.epistola.suite.security.Permission
 import app.epistola.suite.security.RequiresPermission
-import app.epistola.suite.templates.model.ContractVersion
+import app.epistola.suite.templates.contracts.model.ContractVersion
 import org.jdbi.v3.core.Jdbi
 import org.jdbi.v3.core.kotlin.mapTo
 import org.springframework.stereotype.Component
 
 /**
- * Gets the latest published contract version for a template.
- * This is the "current" contract that downstream systems should use.
- * Returns null if no published contract exists.
+ * Gets the latest contract version for a template (draft preferred, then latest published).
  */
-data class GetLatestPublishedContractVersion(
+data class GetLatestContractVersion(
     val templateId: TemplateId,
 ) : Query<ContractVersion?>,
     RequiresPermission {
@@ -25,18 +23,18 @@ data class GetLatestPublishedContractVersion(
 }
 
 @Component
-class GetLatestPublishedContractVersionHandler(
+class GetLatestContractVersionHandler(
     private val jdbi: Jdbi,
-) : QueryHandler<GetLatestPublishedContractVersion, ContractVersion?> {
-    override fun handle(query: GetLatestPublishedContractVersion): ContractVersion? = jdbi.withHandle<ContractVersion?, Exception> { handle ->
+) : QueryHandler<GetLatestContractVersion, ContractVersion?> {
+    override fun handle(query: GetLatestContractVersion): ContractVersion? = jdbi.withHandle<ContractVersion?, Exception> { handle ->
         handle.createQuery(
             """
                 SELECT id, tenant_key, catalog_key, template_key, schema, data_model, data_examples,
                        status, created_at, published_at, created_by
                 FROM contract_versions
                 WHERE tenant_key = :tenantKey AND catalog_key = :catalogKey
-                  AND template_key = :templateKey AND status = 'published'
-                ORDER BY id DESC
+                  AND template_key = :templateKey
+                ORDER BY CASE status WHEN 'draft' THEN 0 ELSE 1 END, id DESC
                 LIMIT 1
                 """,
         )
