+++ modules/epistola-core/src/main/kotlin/app/epistola/suite/templates/contracts/queries/GetLatestPublishedContractVersion.kt
@@ -1,4 +1,4 @@
-package app.epistola.suite.templates.queries.contracts
+package app.epistola.suite.templates.contracts.queries
 
 import app.epistola.suite.common.ids.TemplateId
 import app.epistola.suite.common.ids.TenantKey
@@ -6,40 +6,45 @@ import app.epistola.suite.mediator.Query
 import app.epistola.suite.mediator.QueryHandler
 import app.epistola.suite.security.Permission
 import app.epistola.suite.security.RequiresPermission
-import app.epistola.suite.templates.model.ContractVersionSummary
+import app.epistola.suite.templates.contracts.model.ContractVersion
 import org.jdbi.v3.core.Jdbi
 import org.jdbi.v3.core.kotlin.mapTo
 import org.springframework.stereotype.Component
 
 /**
- * Lists all contract versions for a template, ordered by version number descending.
+ * Gets the latest published contract version for a template.
+ * This is the "current" contract that downstream systems should use.
+ * Returns null if no published contract exists.
  */
-data class ListContractVersions(
+data class GetLatestPublishedContractVersion(
     val templateId: TemplateId,
-) : Query<List<ContractVersionSummary>>,
+) : Query<ContractVersion?>,
     RequiresPermission {
     override val permission: Permission get() = Permission.TEMPLATE_VIEW
     override val tenantKey: TenantKey get() = templateId.tenantKey
 }
 
 @Component
-class ListContractVersionsHandler(
+class GetLatestPublishedContractVersionHandler(
     private val jdbi: Jdbi,
-) : QueryHandler<ListContractVersions, List<ContractVersionSummary>> {
-    override fun handle(query: ListContractVersions): List<ContractVersionSummary> = jdbi.withHandle<List<ContractVersionSummary>, Exception> { handle ->
+) : QueryHandler<GetLatestPublishedContractVersion, ContractVersion?> {
+    override fun handle(query: GetLatestPublishedContractVersion): ContractVersion? = jdbi.withHandle<ContractVersion?, Exception> { handle ->
         handle.createQuery(
             """
-                SELECT id, status, created_at, published_at
+                SELECT id, tenant_key, catalog_key, template_key, schema, data_model, data_examples,
+                       status, created_at, published_at, created_by
                 FROM contract_versions
                 WHERE tenant_key = :tenantKey AND catalog_key = :catalogKey
-                  AND template_key = :templateKey
+                  AND template_key = :templateKey AND status = 'published'
                 ORDER BY id DESC
+                LIMIT 1
                 """,
         )
             .bind("tenantKey", query.templateId.tenantKey)
             .bind("catalogKey", query.templateId.catalogKey)
             .bind("templateKey", query.templateId.key)
-            .mapTo<ContractVersionSummary>()
-            .list()
+            .mapTo<ContractVersion>()
+            .findOne()
+            .orElse(null)
     }
 }
