+++ modules/epistola-core/src/main/kotlin/app/epistola/suite/templates/contracts/queries/ListContractVersions.kt
@@ -1,4 +1,4 @@
-package app.epistola.suite.templates.queries.contracts
+package app.epistola.suite.templates.contracts.queries
 
 import app.epistola.suite.common.ids.TemplateId
 import app.epistola.suite.common.ids.TenantKey
@@ -6,38 +6,40 @@ import app.epistola.suite.mediator.Query
 import app.epistola.suite.mediator.QueryHandler
 import app.epistola.suite.security.Permission
 import app.epistola.suite.security.RequiresPermission
-import app.epistola.suite.templates.model.ContractVersion
+import app.epistola.suite.templates.contracts.model.ContractVersionSummary
 import org.jdbi.v3.core.Jdbi
 import org.jdbi.v3.core.kotlin.mapTo
 import org.springframework.stereotype.Component
 
-data class GetDraftContractVersion(
+/**
+ * Lists all contract versions for a template, ordered by version number descending.
+ */
+data class ListContractVersions(
     val templateId: TemplateId,
-) : Query<ContractVersion?>,
+) : Query<List<ContractVersionSummary>>,
     RequiresPermission {
     override val permission: Permission get() = Permission.TEMPLATE_VIEW
     override val tenantKey: TenantKey get() = templateId.tenantKey
 }
 
 @Component
-class GetDraftContractVersionHandler(
+class ListContractVersionsHandler(
     private val jdbi: Jdbi,
-) : QueryHandler<GetDraftContractVersion, ContractVersion?> {
-    override fun handle(query: GetDraftContractVersion): ContractVersion? = jdbi.withHandle<ContractVersion?, Exception> { handle ->
+) : QueryHandler<ListContractVersions, List<ContractVersionSummary>> {
+    override fun handle(query: ListContractVersions): List<ContractVersionSummary> = jdbi.withHandle<List<ContractVersionSummary>, Exception> { handle ->
         handle.createQuery(
             """
-                SELECT id, tenant_key, catalog_key, template_key, schema, data_model, data_examples,
-                       status, created_at, published_at, created_by
+                SELECT id, status, created_at, published_at
                 FROM contract_versions
                 WHERE tenant_key = :tenantKey AND catalog_key = :catalogKey
-                  AND template_key = :templateKey AND status = 'draft'
+                  AND template_key = :templateKey
+                ORDER BY id DESC
                 """,
         )
             .bind("tenantKey", query.templateId.tenantKey)
             .bind("catalogKey", query.templateId.catalogKey)
             .bind("templateKey", query.templateId.key)
-            .mapTo<ContractVersion>()
-            .findOne()
-            .orElse(null)
+            .mapTo<ContractVersionSummary>()
+            .list()
     }
 }
