+++ modules/epistola-core/src/test/kotlin/app/epistola/suite/templates/contracts/ContractVersionCommandsTest.kt
@@ -14,15 +14,15 @@ import app.epistola.suite.environments.commands.CreateEnvironment
 import app.epistola.suite.mediator.execute
 import app.epistola.suite.mediator.query
 import app.epistola.suite.templates.commands.CreateDocumentTemplate
-import app.epistola.suite.templates.commands.contracts.CreateContractVersion
-import app.epistola.suite.templates.commands.contracts.PublishContractVersion
-import app.epistola.suite.templates.commands.contracts.UpdateContractVersion
 import app.epistola.suite.templates.commands.versions.CreateVersion
 import app.epistola.suite.templates.commands.versions.PublishToEnvironment
-import app.epistola.suite.templates.model.ContractVersionStatus
-import app.epistola.suite.templates.queries.contracts.GetDraftContractVersion
-import app.epistola.suite.templates.queries.contracts.GetLatestContractVersion
-import app.epistola.suite.templates.queries.contracts.ListContractVersions
+import app.epistola.suite.templates.contracts.commands.CreateContractVersion
+import app.epistola.suite.templates.contracts.commands.PublishContractVersion
+import app.epistola.suite.templates.contracts.commands.UpdateContractVersion
+import app.epistola.suite.templates.contracts.model.ContractVersionStatus
+import app.epistola.suite.templates.contracts.queries.GetDraftContractVersion
+import app.epistola.suite.templates.contracts.queries.GetLatestContractVersion
+import app.epistola.suite.templates.contracts.queries.ListContractVersions
 import app.epistola.suite.templates.queries.versions.GetDraft
 import app.epistola.suite.testing.IntegrationTestBase
 import app.epistola.suite.testing.TestIdHelpers
