+++ modules/epistola-core/src/test/kotlin/app/epistola/suite/templates/contracts/SchemaPathNavigatorTest.kt
@@ -1,4 +1,4 @@
-package app.epistola.suite.templates.analysis
+package app.epistola.suite.templates.contracts
 
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.BeforeEach
